pyproject.toml
[project]
name = "aria-travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel==6.30.1",
    "langchain>=1.2.4",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
]


[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

